[SQLLINE-38] !run support ~ expansion from home
