+++ src/main/java/sqlline/OutputFile.java
@@ -24,24 +24,8 @@ public class OutputFile {
   final PrintWriter out;
 
   public OutputFile(String filename) throws IOException {
-    filename = expand(filename);
     file = new File(filename);
     out = new PrintWriter(new FileWriter(file), true);
-  }
-
-  /** Expands "~" to the home directory. */
-  private static String expand(String filename) {
-    if (filename.startsWith("~" + File.separator)) {
-      try {
-        String home = System.getProperty("user.home");
-        if (home != null) {
-          return home + filename.substring(1);
-        }
-      } catch (SecurityException e) {
-        // ignore
-      }
-    }
-    return filename;
   }
 
   @Override public String toString() {
